requirements.txt
fastapi
uvicorn
gspread
pandas
pydantic
python-multipart
requests
pytz
jinja2
itsdangerous
fastapi
uvicorn
gspread
pandas
pydantic
python-multipart
requests
pytz
jinja2
itsdangerous
python-dotenv